# services/api/docs/swagger.yaml
definitions:
...
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
...
  model.CreateUserWirdRequest:
    properties:
      arabic:
        type: string
      count:
        type: integer
      note:
        type: string
      occasion:
        type: string
      source:
        type: string
      title:
        maxLength: 256
        type: string
      translation:
        type: string
      transliteration:
        type: string
    required:
    - title
    type: object
...
  model.UpdateUserWirdRequest:
    properties:
      arabic:
        type: string
      count:
        type: integer
      note:
        type: string
      occasion:
        type: string
      source:
        type: string
      title:
        type: string
      translation:
        type: string
      transliteration:
        type: string
    type: object
...
  service.ZakatFitrahRequest:
    properties:
      person_count:
        minimum: 1
        type: integer
      staple_price_per_kg:
        minimum: 0
        type: number
    required:
    - person_count
    type: object
  service.ZakatMaalRequest:
    properties:
      gold_price_per_gram:
        minimum: 0
        type: number
      total_wealth:
        minimum: 0
        type: number
    type: object
info:
  contact: {}
paths:
...
  /user-wird:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get all user wirds
      tags:
      - Ibadah
      - User Wird
    post:
      consumes:
      - application/json
      parameters:
      - description: User wird request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateUserWirdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Create a user wird
      tags:
      - Ibadah
      - User Wird
  /user-wird/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: User wird ID (uuid)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Delete a user wird by ID
      tags:
      - Ibadah
      - User Wird
    put:
      consumes:
      - application/json
      parameters:
      - description: User wird ID (uuid)
        in: path
        name: id
        required: true
        type: string
      - description: User wird update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUserWirdRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Update a user wird by ID
      tags:
      - Ibadah
      - User Wird
...
  /zakat/fitrah:
    post:
      consumes:
      - application/json
      parameters:
      - description: Zakat fitrah request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ZakatFitrahRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Calculate zakat fitrah
      tags:
      - Ibadah
      - Zakat
  /zakat/maal:
    post:
      consumes:
      - application/json
      parameters:
      - description: Zakat maal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ZakatMaalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Calculate zakat maal (wealth)
      tags:
      - Ibadah
      - Zakat
  /zakat/nishab:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get nishab threshold
      tags:
      - Ibadah
      - Zakat
swagger: "2.0"